Fix bottom border when header is only item in group
We use SettingsList.Header to add extra padding below list items, or to
add explanation text below an item (like the Display Zoom item in iPhone
Settings > Display & Brightness: https://goo.gl/kVLEs5). Currently,
react-native-notifications inserts a group below each header, even if
the group is empty, which adds a border below the header. This commit
only renders a group if there are items in the group to show.